fix(oop): make some chages as requested
