+++ lib/board.rb
@@ -1,7 +1,9 @@
-require "../lib/game_engine"
+require_relative  "../lib/game_engine"
+require_relative '../lib/gui'
 
 class Board
   include GameEngine
+  include GameMessages
   def initialize(p1, p2, d, game_manager)
     @display = d
     @game_manager = game_manager
@@ -28,42 +30,13 @@ class Board
   end
 
   private
-  def game_over?(_name = 'player')
-    @show_player_win = "Player 🏅 #{player_select.name.upcase} 🥇 wins  💰💵 #{player_earnings.to_s.green} 🤑 🏆 ".light_blue
-    @show_game_draw = 'game draw' if @current_turn == 1
-    win_x = ->(n) { n == "#{@player_one.icon} " }
-    win_o = ->(n) { n == "#{@player_two.icon} " }
-    # conditions
-    one = [@slots[:one], @slots[:two], @slots[:three]]
-    two = [@slots[:four], @slots[:five], @slots[:six]]
-    three = [@slots[:seven], @slots[:eight], @slots[:nine]]
-    four = [@slots[:one], @slots[:four], @slots[:seven]]
-    five = [@slots[:two], @slots[:five], @slots[:eight]]
-    six = [@slots[:three], @slots[:six], @slots[:nine]]
-    seven  = [@slots[:one], @slots[:five], @slots[:nine]]
-    eight  = [@slots[:three], @slots[:five], @slots[:seven]]
-
-    case true
-    when one.all?(win_x) || one.all?(win_o) || two.all?(win_x) || two.all?(win_o) ||
-        three.all?(win_x) || three.all?(win_o)
-      @display.msg(@show_player_win)
-      @is_game_over = true
-
-    when four.all?(&win_x) || four.all?(&win_o) || five.all?(&win_x) || five.all?(&win_o)
-      @display.msg(@show_player_win)
-      @is_game_over = true
-
-    when six.all?(&win_x) || six.all?(&win_o) ||
-        seven.all?(&win_x) || seven.all?(&win_o) || eight.all?(&win_x) || eight.all?(&win_o)
-      @display.msg(@show_player_win)
-      @is_game_over = true
-
-    else
-      puts @show_game_draw
-    end
+  def game_over?
+      win_or_draw_check
+      show_winner_msg(player_select.name)
 
     if @is_game_over
-      create_or_append_file([@show_player_win, @display.display_board(@slots), Time.now])
+      create_or_append_file([show_winner_msg(player_select.name),
+                             @display.display_board(@slots), Time.now])
     end
   end
 
@@ -73,12 +46,11 @@ class Board
   end
 
   def can_board_update?
-    @display.msg("Player #{player_select.name} turn to play enter word ranging from one to nine ".green)
+    player_turn_msg(player_select.name)
     player_slot_input = gets.chomp
 
     if !@slots.key?(player_slot_input.to_sym) # checks to see if player input matches any key
-
-      @display.msg('please enter word ranging from one to nine or choose another slot'.red)
+      turn_error_msg
       @update = false
     else
       @slots[player_slot_input.to_sym] = player_select.icon + ' '
@@ -104,6 +76,41 @@ class Board
 
   def player_earnings
      @player_one.bet_amount + @player_two.bet_amount
+  end
+
+
+  def win_or_draw_check
+    win_x = ->(n) { n == "#{@player_one.icon} " }
+    win_o = ->(n) { n == "#{@player_two.icon} " }
+    # conditions
+    one = [@slots[:one], @slots[:two], @slots[:three]]
+    two = [@slots[:four], @slots[:five], @slots[:six]]
+    three = [@slots[:seven], @slots[:eight], @slots[:nine]]
+    four = [@slots[:one], @slots[:four], @slots[:seven]]
+    five = [@slots[:two], @slots[:five], @slots[:eight]]
+    six = [@slots[:three], @slots[:six], @slots[:nine]]
+    seven  = [@slots[:one], @slots[:five], @slots[:nine]]
+    eight  = [@slots[:three], @slots[:five], @slots[:seven]]
+
+    case true
+    when one.all?(win_x) || one.all?(win_o) || two.all?(win_x) || two.all?(win_o) ||
+      three.all?(win_x) || three.all?(win_o)
+      show_winner_msg(player_select.name)
+      @is_game_over = true
+
+    when four.all?(&win_x) || four.all?(&win_o) || five.all?(&win_x) || five.all?(&win_o)
+        @display.msg(@show_player_win)
+        show_winner_msg(player_select.name)
+        @is_game_over = true
+
+    when six.all?(&win_x) || six.all?(&win_o) ||
+        seven.all?(&win_x) || seven.all?(&win_o) || eight.all?(&win_x) || eight.all?(&win_o)
+        show_winner_msg(player_select.name)
+        @is_game_over = true
+
+    else
+      show_draw_msg
+    end
   end
 
 public
